Extract the SSRF-safe avatar fetch into AvatarService
Continues the #318 belt. The avatar fetch/store was three private
controller methods; it moves verbatim into a dedicated AvatarService
(internal sealed), so the controller only decides when to invoke it and
the SSRF-safe transport, content-type allow-list, and size cap live in
one testable unit.

- New AvatarService: TrySetAsync + the private ConnectCallback (per-hop
  IP validation, DNS-rebinding guard) and ReadCappedAsync (size cap),
  moved with every security comment intact.
- The controller constructs it with the same dependencies plus the
  shared User-Agent string passed in, so the outbound-HTTP region the
  controller still owns (OIDC discovery, PKCE fetch) is untouched; the
  now-unused System.Net / System.Net.Sockets usings are dropped, and the
  IProviderManager / IServerConfigurationManager fields (only the avatar
  methods read them) are dropped too - the ctor passes the parameters
  straight to AvatarService.
- Authenticate calls _avatarService.TrySetAsync.
- Adds fail-closed guard-path tests (null URL, and disallowed URLs:
  loopback, cloud-metadata, private range, non-http scheme, unparseable)
  asserting no fetch and no profile-image write - coverage the avatar
  guards lacked as controller privates.

Behavior-preserving: public routes unchanged, 706 existing tests green;
controller loses ~175 lines.

Refs #318
Closes #374
